move readStepValue
